Add failing test for repeated requests just before period
At the moment, the user counter expiry is reset every time a successful
request is made. This means that if a request is repeatedly made just before
the timeout expires, the rate limit can be hit without actually exceeding it.

Add a failing test to show this - not intended for merge as it's very slow,
it's just illustrative.